Introduce shapes
Prior to this change, the functionality of d3-axes was limited to
managing the relationship of one axis to one other axis, i.e. a pair of
axes.

This change introduces `axesShape`, which is capable of wrapping an axes
instance (see d3-wrap), it then automatically scales the x and y
accessors of a path generator before performing the General Update
Pattern.