Demo axesShape
The new axesShape function encapsulates the General Update Pattern for a
path generator and it also handles scaling to the axes' scales. Adding a
curve to the plot is now greatly simplified to a call to axesShape. E.g.

    wrap(axes(...), axesShape(line()).data(...))

An intermediary call to plot is added to render the curve before the
animation, so that the curve is also animated.