Soft dependency on d3-transition
Prior to this change if the selection had a selection method, it was
assumed to be a transition and the transition method was assumed to be
present on the result of the selection call. This means if you provide
something that is transition like, i.e. had a selection method, then the
result of that selection method call must have a transition method.

This change softens the second assumption (i.e. the presence of the
transition method). Therefore, if you provide something that is
semi-transition like then world won't break if you're not using
d3-transition. Note that if you do provide something semi-transition
like and you do use d3-transition, then strange thing might start
happening (i.e. unwanted transitions).