Ignore return value of axes
Prior to this change, the current selection was subsituted for the
return value of the axes component. Most d3 callable components don't
return anything, let alone a selection; this is because
d3-selection#call ignores the return value (call always the selection
for deterministic chaining). This subsitution fails if the axes is
actually wrapped or a composite.

This change ignores the return value of axes in the shape component and,
just to make our intentions clear, the axes component no longer returns
anything.