Avoid selection instance test in d3-gup
Prior to this change, we used the convient `.call(gup, data)` form to
invoke a d3-gup. This uses a fragile selection instance test to convert
the call to the equivalent `.call(gup(data))` form. This can cause
problems if your bundler is faulty (e.g. loads more than on d3-selection
module prototype).

This change avoids the selection instance test by not using the
convenience form.

Also, `axesShape` no longer retains the data, instead it delegates to
d3-gup. This allows us specify a key function with the data.