Split post phase across pre and enter
Prior to this change, update to path and position of axesShape occured
in the post gup phase only. This means that the entering elements share
the same transition as the preexisting elements.

This change splits the update across pre and enter phases so that they
may have custom transitions for each.