Reconcile clipPath's enter and select methods
Prior to this change, clipPath's select method would return a clipPath
element but the enter method would return the rect within it. This means
the attributes (width and height) are not applied when the clipPath is
created.

This change appends the rect inside a call so that the original
selection is returned from enter method.