Recompute and delegate to axesShape's data
Prior to this change, an axesShape's data became stale when it's GUP was
updated and the GUP was called anew when its data was updated.

This change ensures that the axesShape's data is updated whenever its
GUP is updated and when its data is updated this update is simply passed
through to the existing GUP function.